back-outils-photovoltaique/requirements.txt
Django==5.2.3
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
beautifulsoup4==4.12.3
blinker==1.7.0
requests==2.31.0
pillow==10.2.0
django-environ
